+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # NEXT RELEASE
 * Use `@classproperty` decorator to ensure that `YaralyzerConfig.args` exists even when requested outside of a CLI invocation
+* Allow relative paths in `YARALYZER_LOG_DIR` env var
+* Use 256 color system for logs
 
 ### 1.0.13
 * `TRACE` option for `--log-level` argument, `log_trace()` method